requirements.txt
pymongo==4.1.1
python-telegram-bot==21.7
python-telegram-bot[job-queue]==21.7